Board member module completed

+++ admin/modules/boardmembers/controller.php
@@ -1,39 +1,49 @@
-<?php	
-	
-	include_once("modules/boardmembers/service.php");
-	class BoardmembersController{
-		function execute()
-		{
-			$s=new BoardmembersService();
-			//$m=new StudentModel();
-			switch($_REQUEST['action'])
-			{
-				case 'add':
-					if(isset($_POST['full_name']) && isset($_POST['title']) && isset($_POST['content'])){
-							$s->add();
-							$list=$s->getALL();
-							include_once("modules/boardmembers/list.php");
-							return;
-						 }
-					
-					include_once("modules/boardmembers/form.php");
-					break;
-				case 'delete':
-					$s->delete($_REQUEST['id']);
-					$list=$s->getAll();
-					include_once("modules/boardmembers/list.php");
-					break;
-				case 'update':
-				    if(isset($_POST['full_name'] )&& isset($_POST['title']) && isset($_POST['content']))
+<?php
+include_once("modules/boardmembers/service.php");
+include_once("modules/boardmembers/validator.php");
+
+class BoardmembersController{
+			function execute()
+				{
+				$s=new BoardmembersService();
+				
+				switch($_REQUEST['action'])
+				{
+					case 'add':
+
+					$e=$s->add();
+					if(count($e)==0)
 					{
-						$s->update();
+						ob_end_clean();
+						header('Location:index.php?page=boardmembers');
+						exit();
+					}	
+						
+						include_once("modules/boardmembers/form.php");
+						break;
+					case 'delete':
+						$s->delete($_REQUEST['id']);
+						$list=$s->getAll();
+						include_once("modules/boardmembers/list.php");
+						break;
+					case 'update':
+						$e=$s->update();
+						if(count($e)==0)
+					{
+						ob_end_clean();
+						header('Location:index.php?page=boardmembers');
+						exit();
+					}	
+						/*if(isset($_POST['full_name'] )&& isset($_POST['title']) && isset($_POST['content']))
+					{
+						/*$s->update();
 						$list=$s->getAll();
 						include_once("modules/boardmembers/list.php");
 						return;
-					}
+					}*/
+					
 					$f=$s->getById($_REQUEST['id']);
-																
-					include_once("modules/boardmembers/form.php");
+				  include_once("modules/boardmembers/form.php");
 				
 					break;
 					
@@ -45,7 +55,7 @@
 			
 			
 		}
-		
+
 	}
